Initial extension split
